requirements.txt
flake8==4.0.1
flake8-bugbear==22.1.11
flake8-commas==2.1.0
flake8-import-order==0.18.1
flake8-polyfill==1.0.2
flake8-variables-names==0.0.5
gunicorn==20.1.0
pep8-naming==0.12.1
pyyaml==6.0

